requirements.txt
# Core
requests>=2.32.0
pandas>=2.2.2
python-dotenv>=1.0.1

# NLP
transformers>=4.42.0
torch>=2.2.0

# Dashboard
streamlit>=1.36.0
altair>=5.3.0

# Optional OpenAI
openai>=1.35.0

# Utils
tqdm>=4.66.4
